requirements.txt
Django==2.0.2
django-allauth==0.35.0
django-rest-auth==0.9.3
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
mysqlclient==1.3.12
pillow==5.0.0
